refactor: Split evaluator/expressions.rs into modular structure
- Created 4 logical modules by expression type:
  * eval.rs - Main eval() entry point and column references (132 lines)
  * predicates.rs - BETWEEN, LIKE, IN list, POSITION, CAST (185 lines)
  * subqueries.rs - Scalar subquery, IN, EXISTS, quantified comparisons (239 lines)
  * special.rs - CASE expressions and function calls (73 lines)
- Converted expressions.rs to mod.rs with module declarations (9 lines)
- All 186 executor tests passing
- No functional changes, pure refactoring

Reduces evaluator/expressions.rs from 537 lines to ~9 lines mod.rs + 4 focused modules
